Adding TensorData. pulling out more functionality
